+++ CHANGELOG.md
@@ -5,11 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v15.1.7...HEAD)
+### Breaking Changes
+* `session.end()` now returns a `Promise<void>` instead of `void`
+
+### Added
+* [PCM-2114](https://inindca.atlassian.net/browse/PCM-2114) - proxy the reinvite property through for the offer (sdp over xmpp)
+
 # [v15.1.7](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v15.1.6...v15.1.7)
 ### Fixed
 * [PCM-2141]((https://inindca.atlassian.net/browse/PCM-2141) - Sanitize auth token from error responses
 
 # [v15.1.6](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v15.1.5...v15.1.6)
+
 ### Fixed
 * [PCM-2100]((https://inindca.atlassian.net/browse/PCM-2093) - Removed overly noisy data channel message logs.